+++ projects/kiwano/kiwano.vcxproj
@@ -9,9 +9,11 @@
     <ClInclude Include="..\..\src\kiwano\2d\action\ActionWalk.h" />
     <ClInclude Include="..\..\src\kiwano\2d\action\ActionTween.h" />
     <ClInclude Include="..\..\src\kiwano\2d\action\Animation.h" />
-    <ClInclude Include="..\..\src\kiwano\2d\Button.h" />
-    <ClInclude Include="..\..\src\kiwano\2d\Component.h" />
+    <ClInclude Include="..\..\src\kiwano\2d\Actor.h" />
     <ClInclude Include="..\..\src\kiwano\2d\GifSprite.h" />
+    <ClInclude Include="..\..\src\kiwano\base\component\Button.h" />
+    <ClInclude Include="..\..\src\kiwano\base\component\Component.h" />
+    <ClInclude Include="..\..\src\kiwano\base\component\ComponentManager.h" />
     <ClInclude Include="..\..\src\kiwano\base\Director.h" />
     <ClInclude Include="..\..\src\kiwano\base\Module.h" />
     <ClInclude Include="..\..\src\kiwano\base\ObjectBase.h" />
@@ -44,7 +46,6 @@
     <ClInclude Include="..\..\src\kiwano\2d\DebugActor.h" />
     <ClInclude Include="..\..\src\kiwano\2d\ShapeActor.h" />
     <ClInclude Include="..\..\src\kiwano\2d\LayerActor.h" />
-    <ClInclude Include="..\..\src\kiwano\2d\Actor.h" />
     <ClInclude Include="..\..\src\kiwano\2d\Stage.h" />
     <ClInclude Include="..\..\src\kiwano\2d\Sprite.h" />
     <ClInclude Include="..\..\src\kiwano\2d\TextActor.h" />
@@ -115,18 +116,19 @@
     <ClCompile Include="..\..\src\kiwano\2d\action\ActionWalk.cpp" />
     <ClCompile Include="..\..\src\kiwano\2d\action\ActionTween.cpp" />
     <ClCompile Include="..\..\src\kiwano\2d\action\Animation.cpp" />
-    <ClCompile Include="..\..\src\kiwano\2d\Button.cpp" />
+    <ClCompile Include="..\..\src\kiwano\2d\Actor.cpp" />
     <ClCompile Include="..\..\src\kiwano\2d\Canvas.cpp" />
-    <ClCompile Include="..\..\src\kiwano\2d\Component.cpp" />
     <ClCompile Include="..\..\src\kiwano\2d\DebugActor.cpp" />
     <ClCompile Include="..\..\src\kiwano\2d\ShapeActor.cpp" />
     <ClCompile Include="..\..\src\kiwano\2d\GifSprite.cpp" />
     <ClCompile Include="..\..\src\kiwano\2d\LayerActor.cpp" />
-    <ClCompile Include="..\..\src\kiwano\2d\Actor.cpp" />
     <ClCompile Include="..\..\src\kiwano\2d\Stage.cpp" />
     <ClCompile Include="..\..\src\kiwano\2d\Sprite.cpp" />
     <ClCompile Include="..\..\src\kiwano\2d\TextActor.cpp" />
     <ClCompile Include="..\..\src\kiwano\2d\Transition.cpp" />
+    <ClCompile Include="..\..\src\kiwano\base\component\Button.cpp" />
+    <ClCompile Include="..\..\src\kiwano\base\component\Component.cpp" />
+    <ClCompile Include="..\..\src\kiwano\base\component\ComponentManager.cpp" />
     <ClCompile Include="..\..\src\kiwano\base\Director.cpp" />
     <ClCompile Include="..\..\src\kiwano\base\Module.cpp" />
     <ClCompile Include="..\..\src\kiwano\base\ObjectBase.cpp" />
