+++ src/kiwano/2d/action/Action.h
@@ -34,10 +34,6 @@ class ActionScheduler;
 
 KGE_DECLARE_SMART_PTR(Action);
 
-/// \~chinese
-/// @brief 动画列表
-typedef IntrusiveList<ActionPtr> ActionList;
-
 /**
  * \~chinese
  * \defgroup Actions 动画
