+++ src/kiwano/base/component/Component.cpp
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include <kiwano/2d/Component.h>
+#include <kiwano/base/component/Component.h>
 #include <kiwano/2d/Actor.h>
 
 namespace kiwano
